requirements.txt
psycopg2-binary==2.9.11
fastapi==0.135.1
uvicorn==0.42.0
sqlalchemy==2.0.48
python-dotenv==1.2.2
joblib==1.5.3
apscheduler==3.11.2
xgboost==3.2.0
pandas==3.0.1
numpy==2.4.3
scikit-learn==1.8.0
